requirements.txt
dotenv
google-genai
rich